+++ standard/wallets/comparison/jest.config.js
@@ -1,14 +1,13 @@
 module.exports = {
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-    testPathIgnorePatterns: ['/node_modules/', '/dist/'],
-    collectCoverage: false,
-    coverageDirectory: 'coverage',
-    coveragePathIgnorePatterns: ['/node_modules/', '/dist/'],
-    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-    transform: {
-        '^.+\\.ts$': 'ts-jest',
-    },
-    testMatch: ['**/tests/**/*.spec.ts'],
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testPathIgnorePatterns: ['/node_modules/', '/dist/'],
+  collectCoverage: false,
+  coverageDirectory: 'coverage',
+  coveragePathIgnorePatterns: ['/node_modules/', '/dist/'],
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+  transform: {
+    '^.+\\.ts$': 'ts-jest',
+  },
+  testMatch: ['**/tests/**/*.spec.ts'],
 };
-
