+++ standard/wallets/comparison/package.json
@@ -1,28 +1,28 @@
 {
-    "name": "example",
-    "version": "0.0.1",
-    "description": "Automated fee comparison tests for multiple TON wallet implementations",
-    "scripts": {
-        "build": "tsc",
-        "test": "jest",
-        "deploy": "ts-node scripts/deploy.ts",
-        "lint": "prettier --check .",
-        "format": "prettier --write ."
-    },
-    "devDependencies": {
-        "@ton/blueprint": "^0.40.0",
-        "@ton/core": "^0.62.0",
-        "@ton/crypto": "^3.2.0",
-        "@ton/sandbox": "^0.37.2",
-        "@ton/test-utils": "^0.12.0",
-        "@ton/ton": "^15.3.1",
-        "@types/jest": "^29.5.0",
-        "@types/node": "^20.2.5",
-        "jest": "^29.5.0",
-        "prettier": "^3.1.0",
-        "ts-jest": "^29.0.5",
-        "ts-node": "^10.9.1",
-        "typescript": "^5.3.2"
-    },
-    "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
+  "name": "example",
+  "version": "0.0.1",
+  "description": "Automated fee comparison tests for multiple TON wallet implementations",
+  "scripts": {
+    "build": "tsc",
+    "test": "jest",
+    "deploy": "ts-node scripts/deploy.ts",
+    "lint": "prettier --check .",
+    "format": "prettier --write ."
+  },
+  "devDependencies": {
+    "@ton/blueprint": "^0.40.0",
+    "@ton/core": "^0.62.0",
+    "@ton/crypto": "^3.2.0",
+    "@ton/sandbox": "^0.37.2",
+    "@ton/test-utils": "^0.12.0",
+    "@ton/ton": "^15.3.1",
+    "@types/jest": "^29.5.0",
+    "@types/node": "^20.2.5",
+    "jest": "^29.5.0",
+    "prettier": "^3.6.2",
+    "ts-jest": "^29.0.5",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.3.2"
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
