+++ standard/wallets/comparison/utils.ts
@@ -1,7 +1,7 @@
-const getRandom = (min:number, max:number) => {
-    return Math.random() * (max - min) + min;
-}
+const getRandom = (min: number, max: number) => {
+  return Math.random() * (max - min) + min;
+};
 
 export const getRandomInt = (min: number, max: number) => {
-    return Math.round(getRandom(min, max));
-}
+  return Math.round(getRandom(min, max));
+};
